+++ Command/StatusCommand.php
@@ -15,11 +15,14 @@ use Kaliop\eZMigrationBundle\API\Value\MigrationDefinition;
  */
 class StatusCommand extends AbstractCommand
 {
+    const STATUS_INVALID = -1;
+
     protected function configure()
     {
         $this->setName('kaliop:migration:status')
             ->setDescription('View the status of a set of migrations.')
             ->addOption('path', null, InputOption::VALUE_OPTIONAL | InputOption::VALUE_IS_ARRAY, "The directory or file to load the migration definitions from")
+            ->addOption('summary', null, InputOption::VALUE_NONE, "Only print summary information")
             ->setHelp(<<<EOT
 The <info>kaliop:migration:status</info> command displays the status of all available migrations:
 
@@ -71,13 +74,25 @@ EOT
         }
         ksort($index);
 
-        if (count($index) > 50000) {
-            $output->writeln("WARNING: printing the status table might take a while as it contains many rows. Please wait...");
+        if (!$input->getOption('summary')) {
+            if (count($index) > 50000) {
+                $output->writeln("WARNING: printing the status table might take a while as it contains many rows. Please wait...");
+            }
+            $output->writeln("\n <info>==</info> All Migrations\n");
         }
 
-        $output->writeln("\n <info>==</info> All Migrations\n");
-
+        $summary = array(
+            self::STATUS_INVALID => array('Invalid', 0),
+            Migration::STATUS_TODO => array('To do', 0),
+            Migration::STATUS_STARTED => array('Started', 0),
+            Migration::STATUS_DONE => array('Done', 0),
+            Migration::STATUS_SUSPENDED => array('Suspended', 0),
+            Migration::STATUS_FAILED => array('Failed', 0),
+            Migration::STATUS_SKIPPED => array('Skipped', 0),
+            Migration::STATUS_PARTIALLY_DONE => array('Partially done', 0),
+        );
         $data = array();
+
         $i = 1;
         foreach ($index as $name => $value) {
             if (!isset($value['migration'])) {
@@ -85,6 +100,9 @@ EOT
                 $notes = '';
                 if ($migrationDefinition->status != MigrationDefinition::STATUS_PARSED) {
                     $notes = '<error>' . $migrationDefinition->parsingError . '</error>';
+                    $summary[self::STATUS_INVALID][1]++;
+                } else {
+                    $summary[Migration::STATUS_TODO][1]++;
                 }
                 $data[] = array(
                     $i++,
@@ -95,6 +113,15 @@ EOT
                 );
             } else {
                 $migration = $value['migration'];
+
+                if (!isset($summary[$migration->status])) {
+                    $summary[$migration->status] = array($migration->status, 0);
+                }
+                $summary[$migration->status][1]++;
+                if ($input->getOption('summary')) {
+                    continue;
+                }
+
                 switch ($migration->status) {
                     case Migration::STATUS_DONE:
                         $status = '<info>executed</info>';
@@ -145,9 +172,19 @@ EOT
             }
         }
 
+        if ($input->getOption('summary')) {
+            $output->writeln("\n <info>==</info> Migrations Summary\n");
+            // do not print info about the not yet supported case
+            unset($summary[Migration::STATUS_PARTIALLY_DONE]);
+            $data = $summary;
+            $headers = array('Status', 'Count');
+        } else {
+            $headers = array('#', 'Migration', 'Status', 'Executed on', 'Notes');
+        }
+
         $table = $this->getHelperSet()->get('table');
         $table
-            ->setHeaders(array('#', 'Migration', 'Status', 'Executed on', 'Notes'))
+            ->setHeaders($headers)
             ->setRows($data);
         $table->render($output);
     }